requirements.txt
-i https://pypi.org/simple
anyio==4.2.0; python_version >= '3.8'
appnope==0.1.4; python_version >= '3.6'
argon2-cffi==23.1.0; python_version >= '3.7'
argon2-cffi-bindings==21.2.0; python_version >= '3.6'
arrow==1.3.0; python_version >= '3.8'
asgiref==3.7.2; python_version >= '3.7'
asttokens==2.4.1
async-lru==2.0.4; python_version >= '3.8'
attrs==23.2.0; python_version >= '3.7'
babel==2.14.0; python_version >= '3.7'
beautifulsoup4==4.12.3; python_full_version >= '3.6.0'
bleach==6.1.0; python_version >= '3.8'
certifi==2024.2.2; python_version >= '3.6'
cffi==1.16.0; python_version >= '3.8'
charset-normalizer==3.3.2; python_full_version >= '3.7.0'
comm==0.2.1; python_version >= '3.8'
contourpy==1.2.0; python_version >= '3.9'
cycler==0.12.1; python_version >= '3.8'
debugpy==1.8.1; python_version >= '3.8'
decorator==5.1.1; python_version >= '3.5'
defusedxml==0.7.1; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
distlib==0.3.8
django==4.2.10; python_version >= '3.8'
et-xmlfile==1.1.0; python_version >= '3.6'
exceptiongroup==1.2.0; python_version >= '3.7'
executing==2.0.1; python_version >= '3.5'
fastjsonschema==2.19.1
filelock==3.13.1; python_version >= '3.8'
fonttools==4.48.1; python_version >= '3.8'
fqdn==1.5.1; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4' and python_version < '4'
h11==0.14.0; python_version >= '3.7'
httpcore==1.0.3; python_version >= '3.8'
httpx==0.26.0; python_version >= '3.8'
idna==3.6; python_version >= '3.5'
image==1.5.33
importlib-metadata==7.0.1; python_version >= '3.8'
importlib-resources==6.1.1; python_version >= '3.8'
ipykernel==6.29.2; python_version >= '3.8'
ipython==8.18.1; python_version >= '3.9'
ipywidgets==8.1.2; python_version >= '3.7'
isoduration==20.11.0; python_version >= '3.7'
jedi==0.19.1; python_version >= '3.6'
jinja2==3.1.3; python_version >= '3.7'
joblib==1.3.2; python_version >= '3.7'
json5==0.9.14
jsonpointer==2.4; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5, 3.6'
jsonschema[format-nongpl]==4.21.1; python_version >= '3.8'
jsonschema-specifications==2023.12.1; python_version >= '3.8'
jupyter==1.0.0
jupyter-client==8.6.0; python_version >= '3.8'
jupyter-console==6.6.3; python_version >= '3.7'
jupyter-core==5.7.1; python_version >= '3.8'
jupyter-events==0.9.0; python_version >= '3.8'
jupyter-lsp==2.2.2; python_version >= '3.8'
jupyter-server==2.12.5; python_version >= '3.8'
jupyter-server-terminals==0.5.2; python_version >= '3.8'
jupyterlab==4.1.1; python_version >= '3.8'
jupyterlab-pygments==0.3.0; python_version >= '3.8'
jupyterlab-server==2.25.3; python_version >= '3.8'
jupyterlab-widgets==3.0.10; python_version >= '3.7'
kiwisolver==1.4.5; python_version >= '3.7'
markupsafe==2.1.5; python_version >= '3.7'
matplotlib==3.8.2; python_version >= '3.9'
matplotlib-inline==0.1.6; python_version >= '3.5'
mistune==3.0.2; python_version >= '3.7'
nbclient==0.9.0; python_full_version >= '3.8.0'
nbconvert==7.16.0; python_version >= '3.8'
nbformat==5.9.2; python_version >= '3.8'
nest-asyncio==1.6.0; python_version >= '3.5'
notebook==7.1.0; python_version >= '3.8'
notebook-shim==0.2.4; python_version >= '3.7'
numpy==1.26.4; python_version >= '3.9'
opencv-python==4.9.0.80; python_version >= '3.6'
openpyxl==3.1.2; python_version >= '3.6'
overrides==7.7.0; python_version >= '3.6'
packaging==23.2; python_version >= '3.7'
pandas==2.2.0; python_version >= '3.9'
pandocfilters==1.5.1; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
parso==0.8.3; python_version >= '3.6'
pexpect==4.9.0
pillow==10.2.0; python_version >= '3.8'
pipenv==2023.12.1; python_version >= '3.8'
platformdirs==4.2.0; python_version >= '3.8'
prometheus-client==0.20.0; python_version >= '3.8'
prompt-toolkit==3.0.43; python_full_version >= '3.7.0'
psutil==5.9.8; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5'
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==15.0.0; python_version >= '3.8'
pycparser==2.21
pygments==2.17.2; python_version >= '3.7'
pyparsing==3.1.1; python_full_version >= '3.6.8'
python-dateutil==2.8.2; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
python-json-logger==2.0.7; python_version >= '3.6'
pytz==2024.1
pyyaml==6.0.1; python_version >= '3.6'
pyzmq==25.1.2; python_version >= '3.6'
qtconsole==5.5.1; python_version >= '3.8'
qtpy==2.4.1; python_version >= '3.7'
referencing==0.33.0; python_version >= '3.8'
requests==2.31.0; python_version >= '3.7'
rfc3339-validator==0.1.4; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
rfc3986-validator==0.1.1; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
rpds-py==0.18.0; python_version >= '3.8'
scikit-learn==1.4.0; python_version >= '3.9'
scipy==1.12.0; python_version >= '3.9'
send2trash==1.8.2; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
setuptools==69.1.0; python_version >= '3.8'
six==1.16.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
sniffio==1.3.0; python_version >= '3.7'
soupsieve==2.5; python_version >= '3.8'
sqlparse==0.4.4; python_version >= '3.5'
stack-data==0.6.3
terminado==0.18.0; python_version >= '3.8'
threadpoolctl==3.3.0; python_version >= '3.8'
tinycss2==1.2.1; python_version >= '3.7'
tomli==2.0.1; python_version >= '3.7'
tornado==6.4; python_version >= '3.8'
traitlets==5.14.1; python_version >= '3.8'
types-python-dateutil==2.8.19.20240106; python_version >= '3.8'
typing-extensions==4.9.0; python_version >= '3.8'
tzdata==2024.1; python_version >= '2'
uri-template==1.3.0; python_version >= '3.7'
urllib3==2.2.0; python_version >= '3.8'
virtualenv==20.25.0; python_version >= '3.7'
wcwidth==0.2.13
webcolors==1.13; python_version >= '3.7'
webencodings==0.5.1
websocket-client==1.7.0; python_version >= '3.8'
widgetsnbextension==4.0.10; python_version >= '3.7'
zipp==3.17.0; python_version >= '3.8'
